+++ packages/ui/src/components/index.ts
@@ -75,3 +75,5 @@ export * from './BytesSize';
 export * from './BucketAccess';
 export * from './ChartWidget';
 export * from './QRCode';
+export * from './RewardWidget';
+export * from './MobileOverlay';