+++ src/components/HomeLayout/HomeLayout.tsx
@@ -1,11 +1,11 @@
 import { observer } from 'mobx-react-lite';
 import { PropsWithChildren, ReactNode } from 'react';
 import { Button, DiscordIcon, Paper, Stack, styled } from '@developer-console/ui';
-import { AccountDropdown } from '../AccountDropdown';
+import { AnalyticsId } from '@developer-console/analytics';
 
 import { Layout } from '../Layout';
 import { DISCORD_LINK } from '~/constants.ts';
-import { AnalyticsId } from '@developer-console/analytics';
+import { AccountDropdown } from '../AccountDropdown';
 
 export type HomeLayoutProps = PropsWithChildren<{
   rightElement?: ReactNode;
@@ -38,8 +38,10 @@ const HomeLayout = ({ children, rightElement, leftElement }: HomeLayoutProps) =>
     disablePaddings
     fullPage
     headerRight={
-      <Stack direction="row" gap={(theme) => theme.spacing(1)}>
+      <Stack direction="row" spacing={2}>
         <Button
+          variant="outlined"
+          color="secondary"
           href={DISCORD_LINK}
           startIcon={<DiscordIcon />}
           className={AnalyticsId.joinDiscordBtn}